Change Illumination Wand to toggle light sources
If the wand would place a light where one already exists, the light is
removed at no cost. If the light's color is different from the wand's,
the light's color will change instead of being removed. This is
consistent with the way shift-using the wand will toggle the block to
and from a TranslucentBlock.