Add compatibility with Lootr.
Lootr provides instanced loot chests for every player on the server. If
a player opens a chest they've never opened before, it generates an
inventory for them and stores it in WorldSavedData.

It primarily helps with multiplayer servers: people who are looking for
loot don't go "oh empty chests/chest with rubbish in it" and then
generating a bunch of chunks looking for loot.

Functionally, the only difference with this change to use
`LockableLootTileEntity.setLootTable` (static version) and the
previously used method is that it doesn't use your specific utility for
fetching the tile entity.

On the plus side, it means that the hook (which can be mixed-in to) has
the relevant world and information allowing for replacement of chests
during generation without having to rely on externally snychronised (and
buggy) ticking lists to try replacing after the chunk is generated.